feat(EMS-571): shared validation checks and error message for requested cover start date instances in single and multiple policy pages
