+++ src/ui/server/content-strings/error-messages.ts
@@ -90,12 +90,12 @@ export const ERROR_MESSAGES = {
           IS_EMPTY: 'Enter the credit period you have with your buyer',
           ABOVE_MAXIMUM: 'The credit period you have with your buyer cannot be more than 1000 characters.',
         },
+        [FIELD_IDS.INSURANCE.POLICY_AND_EXPORTS.CONTRACT_POLICY.REQUESTED_START_DATE]: {
+          IS_EMPTY: 'Enter a policy start date in the correct format - for example, 06 11 2023',
+          NOT_A_NUMBER: 'Enter a policy start date in the correct format - for example, 06 11 2023',
+          BEFORE_EARLIEST: 'You cannot enter a policy start date in the past - enter a future date',
+        },
         SINGLE: {
-          [FIELD_IDS.INSURANCE.POLICY_AND_EXPORTS.CONTRACT_POLICY.REQUESTED_START_DATE]: {
-            IS_EMPTY: 'Enter a policy start date in the correct format - for example, 06 11 2023',
-            NOT_A_NUMBER: 'Enter a policy start date in the correct format - for example, 06 11 2023',
-            BEFORE_EARLIEST: 'You cannot enter a policy start date in the past - enter a future date',
-          },
           [FIELD_IDS.INSURANCE.POLICY_AND_EXPORTS.CONTRACT_POLICY.SINGLE.TOTAL_CONTRACT_VALUE]: {
             IS_EMPTY: 'Enter your contract value as a whole number - do not enter decimals',
             NOT_A_NUMBER: 'Enter your contract value as a whole number - do not enter decimals',
@@ -104,13 +104,7 @@ export const ERROR_MESSAGES = {
             ABOVE_MAXIMUM: 'The maximum the buyer will owe cannot be more than £499.999. If you need cover for more than this, fill in this form instead.',
           },
         },
-        MULTIPLE: {
-          [FIELD_IDS.INSURANCE.POLICY_AND_EXPORTS.CONTRACT_POLICY.REQUESTED_START_DATE]: {
-            IS_EMPTY: 'Enter a policy start date in the correct format - for example, 06 11 2023',
-            NOT_A_NUMBER: 'Enter a policy start date in the correct format - for example, 06 11 2023',
-            BEFORE_EARLIEST: 'You cannot enter a policy start date in the past - enter a future date',
-          },
-        },
+        MULTIPLE: {},
       },
     },
   },
